Refactor trigger utilities and add payload support
Move GitHub trigger utilities to dedicated file and enhance
resolveTrigger to support direct payload processing
