+++ packages/giselle-engine/src/core/flows/resolve-trigger.ts
@@ -1,6 +1,7 @@
 import {
 	type CompletedGeneration,
 	GenerationContext,
+	type GenerationInput,
 	type GenerationOutput,
 	type QueuedGeneration,
 	isTriggerNode,
@@ -11,12 +12,18 @@ import {
 	setGenerationIndex,
 	setNodeGenerationIndex,
 } from "../generations/utils";
+import type { GitHubEvent } from "../github/events";
+import {
+	buildTriggerInputs,
+	parseCommandFromEvent,
+} from "../github/trigger-utils";
 import type { GiselleEngineContext } from "../types";
 import { getFlowTrigger } from "./utils";
 
 export async function resolveTrigger(args: {
 	context: GiselleEngineContext;
 	generation: QueuedGeneration;
+	payload?: unknown;
 }) {
 	const operationNode = args.generation.context.operationNode;
 	if (!isTriggerNode(operationNode)) {
@@ -32,14 +39,25 @@ export async function resolveTrigger(args: {
 
 	const generationContext = GenerationContext.parse(args.generation.context);
 
+	let payloadInputs: GenerationInput[] | null = null;
+	if (
+		args.payload !== undefined &&
+		triggerData.configuration.provider === "github"
+	) {
+		payloadInputs = buildTriggerInputs({
+			githubTrigger: githubTriggers[triggerData.configuration.event.id],
+			trigger: triggerData,
+			githubEvent: args.payload as GitHubEvent,
+		});
+	}
+
 	const outputs: GenerationOutput[] = [];
 	switch (triggerData.configuration.provider) {
 		case "github": {
 			const trigger = githubTriggers[triggerData.configuration.event.id];
+			const inputsToUse = payloadInputs ?? generationContext.inputs ?? [];
 			for (const payload of trigger.event.payloads.keyof().options) {
-				const input = generationContext.inputs?.find(
-					(input) => input.name === payload,
-				);
+				const input = inputsToUse.find((i) => i.name === payload);
 				if (input === undefined) {
 					continue;
 				}
