+++ packages/giselle-engine/src/core/github/handle-webhook.ts
@@ -244,272 +244,6 @@ async function addReaction(args: {
 	}
 }
 
-function buildTriggerInputs(args: {
-	githubTrigger: (typeof githubTriggers)[keyof typeof githubTriggers];
-	trigger: Awaited<ReturnType<typeof getFlowTrigger>>;
-	githubEvent: GitHubEvent;
-}): GenerationInput[] | null {
-	const { githubTrigger, githubEvent, trigger } = args;
-	switch (githubTrigger.event.id) {
-		case "github.issue.created":
-			return buildIssueCreatedInputs(
-				githubEvent,
-				githubTrigger.event.payloads.keyof().options,
-			);
-		case "github.issue.closed":
-			return buildIssueClosedInputs(
-				githubEvent,
-				githubTrigger.event.payloads.keyof().options,
-			);
-		case "github.issue_comment.created":
-			if (trigger.configuration.event.id !== "github.issue_comment.created") {
-				return null;
-			}
-			return buildIssueCommentInputs(
-				githubEvent,
-				githubTrigger.event.payloads.keyof().options,
-				trigger.configuration.event.conditions.callsign,
-			);
-		case "github.pull_request_comment.created":
-			if (
-				trigger.configuration.event.id !== "github.pull_request_comment.created"
-			) {
-				return null;
-			}
-			return buildIssueCommentInputs(
-				githubEvent,
-				githubTrigger.event.payloads.keyof().options,
-				trigger.configuration.event.conditions.callsign,
-			);
-		case "github.pull_request.opened": {
-			if (githubEvent.type !== GitHubEventType.PULL_REQUEST_OPENED) {
-				return null;
-			}
-			const triggerInputs: GenerationInput[] = [];
-			for (const payload of githubTrigger.event.payloads.keyof().options) {
-				switch (payload) {
-					case "title":
-						triggerInputs.push({
-							name: "title",
-							value: githubEvent.payload.pull_request.title,
-						});
-						break;
-					case "body":
-						triggerInputs.push({
-							name: "body",
-							value: githubEvent.payload.pull_request.body ?? "",
-						});
-						break;
-					case "number":
-						triggerInputs.push({
-							name: "number",
-							value: githubEvent.payload.pull_request.number.toString(),
-						});
-						break;
-					case "pullRequestUrl":
-						triggerInputs.push({
-							name: "pullRequesturl",
-							value: githubEvent.payload.pull_request.html_url,
-						});
-						break;
-					default: {
-						const _exhaustiveCheck: never = payload;
-						throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-					}
-				}
-			}
-			return triggerInputs;
-		}
-		case "github.pull_request.ready_for_review": {
-			if (githubEvent.type !== GitHubEventType.PULL_REQUEST_READY_FOR_REVIEW) {
-				return null;
-			}
-			const triggerInputs: GenerationInput[] = [];
-			for (const payload of githubTrigger.event.payloads.keyof().options) {
-				switch (payload) {
-					case "title":
-						triggerInputs.push({
-							name: "title",
-							value: githubEvent.payload.pull_request.title,
-						});
-						break;
-					case "body":
-						triggerInputs.push({
-							name: "body",
-							value: githubEvent.payload.pull_request.body ?? "",
-						});
-						break;
-					case "number":
-						triggerInputs.push({
-							name: "number",
-							value: githubEvent.payload.pull_request.number.toString(),
-						});
-						break;
-					case "pullRequestUrl":
-						triggerInputs.push({
-							name: "pullRequestUrl",
-							value: githubEvent.payload.pull_request.html_url,
-						});
-						break;
-					default: {
-						const _exhaustiveCheck: never = payload;
-						throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-					}
-				}
-			}
-			return triggerInputs;
-		}
-		case "github.pull_request.closed": {
-			if (githubEvent.type !== GitHubEventType.PULL_REQUEST_CLOSED) {
-				return null;
-			}
-			const triggerInputs: GenerationInput[] = [];
-			for (const payload of githubTrigger.event.payloads.keyof().options) {
-				switch (payload) {
-					case "title":
-						triggerInputs.push({
-							name: "title",
-							value: githubEvent.payload.pull_request.title,
-						});
-						break;
-					case "body":
-						triggerInputs.push({
-							name: "body",
-							value: githubEvent.payload.pull_request.body ?? "",
-						});
-						break;
-					case "number":
-						triggerInputs.push({
-							name: "number",
-							value: githubEvent.payload.pull_request.number.toString(),
-						});
-						break;
-					case "pullRequestUrl":
-						triggerInputs.push({
-							name: "pullRequestUrl",
-							value: githubEvent.payload.pull_request.html_url,
-						});
-						break;
-					default: {
-						const _exhaustiveCheck: never = payload;
-						throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-					}
-				}
-			}
-			return triggerInputs;
-		}
-		default: {
-			const _exhaustiveCheck: never = githubTrigger.event;
-			throw new Error(`Unhandled event id: ${_exhaustiveCheck}`);
-		}
-	}
-}
-
-function buildIssueCreatedInputs(
-	githubEvent: GitHubEvent,
-	payloads: readonly ("title" | "body")[],
-): GenerationInput[] | null {
-	if (githubEvent.type !== GitHubEventType.ISSUES_OPENED) {
-		return null;
-	}
-
-	const inputs: GenerationInput[] = [];
-	for (const payload of payloads) {
-		switch (payload) {
-			case "title":
-				inputs.push({ name: "title", value: githubEvent.payload.issue.title });
-				break;
-			case "body":
-				inputs.push({
-					name: "body",
-					value: githubEvent.payload.issue.body ?? "",
-				});
-				break;
-			default: {
-				const _exhaustiveCheck: never = payload;
-				throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-			}
-		}
-	}
-	return inputs;
-}
-
-function buildIssueClosedInputs(
-	githubEvent: GitHubEvent,
-	payloads: readonly ("title" | "body")[],
-): GenerationInput[] | null {
-	if (githubEvent.type !== GitHubEventType.ISSUES_CLOSED) {
-		return null;
-	}
-
-	const inputs: GenerationInput[] = [];
-	for (const payload of payloads) {
-		switch (payload) {
-			case "title":
-				inputs.push({ name: "title", value: githubEvent.payload.issue.title });
-				break;
-			case "body":
-				inputs.push({
-					name: "body",
-					value: githubEvent.payload.issue.body ?? "",
-				});
-				break;
-			default: {
-				const _exhaustiveCheck: never = payload;
-				throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-			}
-		}
-	}
-	return inputs;
-}
-
-function buildIssueCommentInputs(
-	githubEvent: GitHubEvent,
-	payloads: readonly ("body" | "issueBody" | "issueNumber" | "issueTitle")[],
-	callsign: string,
-): GenerationInput[] | null {
-	if (githubEvent.type !== GitHubEventType.ISSUE_COMMENT_CREATED) {
-		return null;
-	}
-
-	const command = parseCommandFromEvent(githubEvent);
-	if (command === null || command.callsign !== callsign) {
-		return null;
-	}
-
-	const inputs: GenerationInput[] = [];
-	for (const payload of payloads) {
-		switch (payload) {
-			case "body":
-				inputs.push({ name: "body", value: command.content });
-				break;
-			case "issueBody":
-				inputs.push({
-					name: "issueBody",
-					value: githubEvent.payload.issue.body ?? "",
-				});
-				break;
-			case "issueNumber":
-				inputs.push({
-					name: "issueNumber",
-					value: githubEvent.payload.issue.number.toString(),
-				});
-				break;
-			case "issueTitle":
-				inputs.push({
-					name: "issueTitle",
-					value: githubEvent.payload.issue.title,
-				});
-				break;
-			default: {
-				const _exhaustiveCheck: never = payload;
-				throw new Error(`Unhandled payload id: ${_exhaustiveCheck}`);
-			}
-		}
-	}
-	return inputs;
-}
-
 // Extracted for legacy parallel execution.
 // @deprecated This function will be removed in the future. Please use the new webhook processing method.
 async function processMatchedIntegrationSettingsDeprecated(
